Bot Selection Menu

+++ src/core/checkers/checkersminmax.hpp
@@ -39,7 +39,7 @@ namespace NETWORK
         std::pair<int32_t, BitMove> minimax(BoardStruct& board_state, std::vector<NTARS::DATA::TrainingData<std::vector<float>>>& trainingData,
              bool max = true, uint32_t currentDepth = 0, uint32_t maxDepth = 1, int32_t alpha = std::numeric_limits<int32_t>::min(), int32_t beta = std::numeric_limits<int32_t>::max());
         int32_t evaluatePosition(BoardStruct& currentBoard, bool max);
-        std::vector<float> getTrainingLabel(uint32_t moveIndex);
+        std::vector<float> getTrainingLabel(uint64_t moveIndex);
         
         int32_t valueMove(BoardStruct& board_state, const BitMove& move, const bool max);
